requirements.txt
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai==0.3.16
llama-index-llms-gemini
llama-index-utils-workflow
llama-parse
python-dotenv
streamlit